BUG: Change the default precision for printing double value
The conversion of a floating-point to text and back is exact as long as
at least `max_digits10` is used (17 for double).

The net effect is that the text representation is now closer to the
DICOM Image Orientation (Patient) or Image Position (Patient) attribute
and will allow easier conversion from one representation to the other.

Since commit e532c682930f ITK explicitly set the default precision to 6,
change of default will be transparent.